+++ src/plugins/blocksmith/components/video-block-modal/index.jsx
@@ -188,7 +188,6 @@ class VideoBlockModal extends React.Component {
 
   handleSelectImage(imageNode) {
     this.setState({
-      autoplay: false,
       selectedImageNode: imageNode,
     });
   }