+++ src/plugins/raven/actions/sendHeartbeat.js
@@ -2,6 +2,8 @@
 import swal from 'sweetalert2';
 import NodeRef from '@gdbots/schemas/gdbots/ncr/NodeRef';
 import getAccessToken from '@triniti/cms/plugins/iam/selectors/getAccessToken';
+import getNode from '@triniti/cms/plugins/ncr/selectors/getNode';
+import hasNode from '@triniti/cms/plugins/ncr/selectors/hasNode';
 import isJwtExpired from '@triniti/cms/plugins/iam/utils/isJwtExpired';
 import { actionTypes, ravenTypes } from '../constants';
 import publishMessage from './publishMessage';
@@ -24,6 +26,16 @@ export default (topic, etag = null) => async (dispatch, getState) => {
       });
       const data = await response.json();
 
+      let username = 'an unknown user or process';
+      if (data.updater_ref) {
+        const userRef = NodeRef.fromString(data.updater_ref);
+
+        if (hasNode(state, userRef)) {
+          const user = getNode(state, userRef);
+          username = user.get('title', `${user.get('first_name')} ${user.get('last_name')}`);
+        }
+      }
+
       if (etag
         && data.ok
         && data.etag !== etag
@@ -37,7 +49,7 @@ export default (topic, etag = null) => async (dispatch, getState) => {
       ) {
         const nodeRef = NodeRef.fromString(topic);
         await swal.fire({
-          html: `This ${nodeRef.getLabel()} has been changed by another person or process.<br/>If you save, you may overwrite their changes.`,
+          html: `This ${nodeRef.getLabel()} has been changed by ${username}.<br/>If you save, you may overwrite their changes.`,
           position: 'top-end',
           showCloseButton: true,
           showConfirmButton: false,